pyproject.toml
[tool.poetry]
name = "todo-api-django"
version = "1.0.0"
description = "A ToDo API created in Python and Django Rest Framework for case study."
authors = ["Bruno Santos <bruno.dos1997@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.6.0"
pre-commit = "^2.20.0"
commitizen = "^2.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
